fix(controls): widen Window drag/move delta arithmetic to long
HandlePointerDrag's move branch computed drag delta and the new window
position with raw int arithmetic that could overflow before Math.Clamp
ever ran. The sibling resize branch in the same method was already
hardened for this in 6715ac8a; mirror that pattern for move by
widening deltaX/deltaY to long before any arithmetic and casting the
clamped result back to int.

Adds a regression test alongside the existing resize extreme-coordinate
test. Because every Length in this codebase is non-negative and a
title-bar Press must land within the window's own absolute bounds, the
divergent-baseline overflow condition cannot be constructed purely
through the public Pointer gesture API (unlike resize's independent
Width baseline) - so the test drives a real Press/Move gesture and
reflectively substitutes _dragWindowOrigin with a value the Press path
could never legitimately produce, to exercise the same hardening.

Fixes #1081

+++ src/SharpVision/Windows/Window.cs
@@ -1066,8 +1066,8 @@ public class Window: FloatingSurfaceBase, IOverlayPositionConstraint
         else if (_dragging)
         {
             var gestureParent = Parent;
-            var deltaX = cells.X - _dragPointerOrigin.X;
-            var deltaY = cells.Y - _dragPointerOrigin.Y;
+            var deltaX = (long) cells.X - _dragPointerOrigin.X;
+            var deltaY = (long) cells.Y - _dragPointerOrigin.Y;
             var clientBounds = Parent?.ContentBounds ?? default;
             var maximumLeft = Math.Max(0, clientBounds.Width - LocalBounds.Width);
             var maximumTop = Math.Max(0, clientBounds.Height - LocalBounds.Height);
@@ -1097,8 +1097,8 @@ public class Window: FloatingSurfaceBase, IOverlayPositionConstraint
 
             Overlay.SetRight(this, null);
             Overlay.SetBottom(this, null);
-            Overlay.SetLeft(this, Length.Cells(Math.Clamp(_dragWindowOrigin.X + deltaX, 0, maximumLeft)));
-            Overlay.SetTop(this, Length.Cells(Math.Clamp(_dragWindowOrigin.Y + deltaY, 0, maximumTop)));
+            Overlay.SetLeft(this, Length.Cells((int) Math.Clamp(_dragWindowOrigin.X + deltaX, 0, maximumLeft)));
+            Overlay.SetTop(this, Length.Cells((int) Math.Clamp(_dragWindowOrigin.Y + deltaY, 0, maximumTop)));
             eventArgs.IsHandled = true;
         }
     }
